# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'TiendaBackend API Documentation'
  description: 'Backend API del sistema de ventas Tienda. Módulo de autenticación y gestión de sesiones.'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtén tu token haciendo <b>POST /api/login</b> con tus credenciales. Luego pégalo aquí para probar los endpoints protegidos.'
security:
  -
    default: []
paths:
...
  /api/products/query:
    post:
      summary: 'Query products (POST)'
      operationId: queryProductsPOST
      description: "Supports two payload formats. Formato B (standard):\n`p.page` = row offset (0-indexed), `p.per_page`, `o.column`, `o.direction`, `w` = object.\nFormato A (compact):\n`p.r` = row offset, `p.s` = per_page, `o.field`, `o.type`, `w` = array of `{f, ao, v, lo}`."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  ok: true
                  code: 200
                  status: OK
                  message: 'Products retrieved.'
                  data:
                    items: []
                    totalCount: 0
                    summary:
                      - 0
                properties:
                  ok:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: OK
                  message:
                    type: string
                    example: 'Products retrieved.'
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        example: []
                      totalCount:
                        type: integer
                        example: 0
                      summary:
                        type: array
                        example:
                          - 0
                        items:
                          type: integer
      tags:
        - Inventory
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                p:
                  type: object
                  description: 'Pagination. `page`/`per_page` or `r`/`s`.'
                  example: []
                  properties: {  }
                f:
                  type: array
                  description: 'Fields to return.'
                  example:
                    - id
                    - name
                  items:
                    type: string
                o:
                  type: object
                  description: 'Order. `column`+`direction` or `field`+`type`.'
                  example: []
                  properties: {  }
                w:
                  type: object|array
                  description: 'Where filters. Object for simple equality, array for advanced operators.'
                  example: architecto
                totalCount:
                  type: boolean
                  description: 'Include total count (default true).'
                  example: true
...